requirements.txt
matplotlib==3.10.7
numpy==2.3.4
opencv_python==4.12.0.88
optuna==4.5.0
Pillow==12.0.0
scikit_learn==1.7.2
seaborn==0.13.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
scikit-image
